smart-sync-attendance/requirements.txt
gradio==4.38.0
deepface
pandas
numpy
opencv-python
Pillow
tf-keras
